update readme, move layout folder
