+++ src/hooks/useEpisodes.ts
@@ -28,7 +28,7 @@ export const useEpisodes = () => {
 
   const genStorePayload = useCallback(
     /*
-      Takes a list of episodes, checks are stored within the local storage
+      Takes a list of episodes, checks if are stored within the local storage
       then saves them if they're not stored. (Updates or init the store)
     */
     (data: IEpisode[]) => {
